Cherrypick f19e45 from master to r15. do not merge.
Logcat autoscroll behavior update.

Rather than checking the position of the thumb every time entries
are added, we listen to the scroll bar being moved up or down by
the user.

In addition, also update the scrolling behavior whenever entries in
the table are selected. Auto scrolling will stop if anything but the
last entry is selected.

Change-Id: I97887c8c3707218d01ed487e39cd71360ded7632