Chery-pick 2c7c729 from master to tools_r15. do not merge.
Fix refresh issues after a library build.

At the end of a library build we need to refresh the library
bin folder to ensure that the main project is notified of the changes.

However it looks like the refresh doesn't trigger a new build unless
it is asynchronous from a different job.

Also fixed the delta visitor used by the main project to look at
libraries to support looking into bin/res to detect output of the
crunch cache.

(Also added some debug output that is currently disabled).

Change-Id: Id2013c608abcda6fda6b0ce4572116edc2502bcb