Fix issue when a project and its libraries use the same jar files. do not merge.
This is only an issue in Ant because in Eclipse we don't automatically
pull the jar files from libraries into the main project (we should somehow
now that we have the Library Projects jar container that is dynamic).

Right now we do a simple size/sha1 check on libraries that have the same
name to figure out if they are the same version. If they are we only
use one in the dex step (that notoriously fails to add the same class
twice). If they are different we stop the build as it's an error (having
two library projects depending on two different versions of a jar file
should be an error as we can be sure the two versions are API compatible).

For later: not use the file name only? find a way to version the libraries
and to have them declare whether they are API compatible with older versions?

Also added a hard-coded case for the Android Support Library. If both the v4
and the v13 are detected, use the v13 only as it includes the v4 already.

New test apps. Three cases:
- main and library projects with duplicate jar files that are identical
- main and library projects with duplicate jar files that are NOT identical
- main and library projects with v4 and v13 in the dependency list.

(cherry picked from commit 44a21c237072fba5454732c74d1f7653523c4105)

Change-Id: I0a159db5e4a412665d3561a66f4ddbc8fe61842d