Never add R/R$* classes to the library jar files. do not merge.
Previous change added excluding Manifest and BuildConfig but
made it so that it was only for the current app package.

The problem is that project that depends on libraries also generates
R classes for their libraries.
If the project itself is a library then the R classes from the libraries
would get in the library jar output.

If the main project had some diamond dependencies in its libraries
then the same R class from a library could be added twice to its
dex input causing an error.

(cherry picked from commit f369217000a89786369810a177e3404f103c4874)

Change-Id: I8fe5abb66e3752638cbb0fa8cec9df7c4d807f19