Fix configuration information for library projects. do not merge.
This changeset fixes one more bug related to lint configurations and
library projects (see issue 26029).

Some lint checks, such as the MergeRootFrameLayoutDetector, computes
the warnings at the end of processing all files. At that point, the
context points to the master project, so any errors which were
actually found in a library project will instead be using the master
project's configuration. That means that any suppress rules applies to
the lint.xml in the library project will be ignored.

This changeset fixes this by moving the logic which looks up the
severity for a warning out of the lint clients and into the
context. Now it checks the current projects being scanned and looks up
the corresponding project for each file (based on the file prefix),
and retrieves the configuration that way.

This changeset also makes one more fix: It now consults *both* the
library project *and* the master project to see if a rule should be
ignored. This means that if you turn off a given check in your master
project, you will no longer see those warnings from library projects
either, which seems desirable.

(cherry picked from commit fd9fe37e806b8f14b2944d29f307cd050ecb419c)

Change-Id: Ia8136609ebee046cdf07999ce93e8b9229f22df2