Close the dependency file after reading it. do not merge.
The dependency file will be eventually closed by the finalize() method
at some point, but it is not known when.  In the meantime, the Ant
recipe continues execution and may try to move or delete this file.
At that point the build may fail because the file is still open.

The solution is to explicitly close the BufferedReader opened by
DependencyGraph.readFile(), so the underlying file is closed as soon
as reading is finished.

Signed-off-by: Kaloian Doganov <doganov@projectoria.bg>

(cherry picked from commit 763b15fd4fd4cf92b6aef2e29de98008894413a5)

Change-Id: I57eba1051be79e8e4f132bfae4573dc18430ea3d