Fix issue with automatic jar dependency. do not merge.
The previous commit relied on a cached version of the Android container
to add jar in to the list. This properly look for inside the libs folder
for jar file, and restrict looking at project classpathes to only
Java-only projects.

Also hooked into the file change listener of the Sdk class to
update the project container when a file is added to removed
to a libs folder.

(cherry picked from commit 906fec2a94663ad1aa76aad55fdf2cbb97324d66)

Change-Id: Ie765303f91abe7df89edaa957506ee8bb8dc9cf0