Handle the android support gridlayout library automatically. do not merge.
This changeset adds support for the android support library's
GridLayout library project.

When you create a new layout with the GridLayout, or when you drop a
GridLayout, the IDE checks whether you need the compatibility version
of GridLayout (e.g. min sdk < 14), and if so, offers to install it.

This will then first run the SDK manager to install the android
support package into extras, and then it creates a local library
project in the Eclipse workspace, and updates the library dependency
to reference it.

Finally, it rewrites tags such that the layout will use the
compatibility package for the <GridLayout> and <Space> tags.  This is
done in the node handler, so client rule code will automatically get
the right compatibility tag; they don't need to handle it there.

(cherry picked from commit 9bfb05e9d9ea4b2b969e50c3096e2fdb95653648)

Change-Id: I2deef0feceb20af17ec7600ac98e1b9f91bde444