SDK Manager: support platform previews. do not merge.
This allows the SDK Manager to create a group for
platform previews (api + codename) which is separate
from the api it overrides.

There's a latent issue with the fact that addons do not
support codenames. This should only happen on nightly
builds and not on officially released platforms though.

(cherry picked from commit 53965f3f406f5a142b71753a4623a4d363795527)

Change-Id: I652bb73270019b56ded2d6954ccfa5f1d78ab342