Ant build now sanitize jar files in setup task. do not merge.
Previously the list of jar files was sanitized (to remove
duplicates) in the dex task, but this meant the full list
(with duplicates) was passed to proguard when building in release
mode.

This changeset move the sanitization of the jar files in the
Setup Task so that the script later only deals with a sanitized
list. The means the content of libs/*.jar for the current project
must be looked at in the task instead of later in the XML script.

(cherry picked from commit a08befd52438d523cb2cd41b53f961785d3872fd)

Change-Id: I0744d9ced05c363c58da788c4bef7759fec79b2c