ADT fix: support Compatibility library in NPW
This fixes the NPW so that it supports the older
"Compatibility" library, not just the new "Support"
library. One issue is that the NPW prompts users to
update, but the code only checks for the support
lib in the new folder (SDK/extras/android/support).
During an upgrade the folder isn't changed and the
code fails to find it at the android/compatibility
location.

SDK Bug: 33859 and duplicates

(cherry picked from commit 848e872d629837043478c7b7b1510c462fb28980)

Change-Id: Icf1a32f320acbc2e1caff92a6d9741aaa5e446a9