Fix editor startup exception for new configurations
If you use the wizard to create a new layout in a new configuration
folder (for example, pick a brand new language qualifier on the second
page of the wizard), then the editor comes up blank.

The reason this happens is that during editor initialization, it
attempts to look up the ResourceFolder, and gets back null which it
then proceeds to dereference.

This fix works around this by simply looking up the configuration
folder directly rather than from the ResourceFolder, since that's all
it's needed for at this point.

This probably fixes
http://code.google.com/p/android/issues/detail?id=24578

(cherry picked from commit a6a51d22457f4ae6a43db992c4712b3a8bfd0a77)

Change-Id: I4b74ae18224b693bc3c88e75b91e6f2d0bac74a4