Make welcome wizard a tiny bit configurable.
The welcome wizard has two steps right now: installing an SDK,
and opting in to data collection. This change allows the
constructor of the WelcomeWizard to select which among those
pages ought to be shown. This is useful in the case of the ADT
package installation where we know there is an SDK, and hence
we don't want to show the installation page.

(cherry picked from commit 41fe9db125c6867bbcff863f39b67cbdfc7e2613)

Change-Id: I61102c156aa3802c808030abb614d929e5ee72be