Merge "SDK Manager: handle platform-tools preview + final."
One issue is that when there was only one instance of platform-tools possible,
the computeUpdates() code would pick the first one. But now there can be 2 of
them (preview, non-preview) and thus we need to pick up the higher one even if
it's not the first choice.

Same issue with tools: if a platform depends on tools and there are none
installed, we need to pick the highest version of the available preview
or final package that satisfies the dependency.

Note that in both cases the issue does not arise if there's already
a tools or platform-tools installed.

(cherry picked from commit ce38196f8e47fa3b48805b36eaca55716c0af79c)

Change-Id: Iec36d68378ac1235107183a2f403e76e419042dc