Merge "Fix refresh issue in the AVD Manager."
In tools 21.0.0, hitting refrehs in the AVD Manager will
create the list display to blink rapidly several times
in a row. Same happens when a new AVD is created from a
device.

The issue is that the DeviceManager reparses and rebuilds
its internal lists from scratch for every single call to
getDevices() and getDeviceStatus(). Each time it notifies
the listeners, and consequently the AVD Manager rebuilds
its table. 2 fixes for that:
- First use a boolean guard to prevent recursive refreshes
  of the avd manager table in AvdSelector.
- Second fix DeviceManager to not rebuilt its lists all
  the time.

This also changes DeviceManager so that callers create
one instance that contains all the lists and there no static
data shared between the instances.
This is more deterministic. It shifts the responsibility to
the callers to pass around the same instance if they want the
data to remain consistent.

(cherry picked from commit 8e151ab957b82e8d0573ac0768f3bd4424caaf30)

Change-Id: I1c55a2fd42b3c2cc37f29294a28476feb3dba372