Merge "SDK Manager: auto-select Support Library at startup."
When the SDK Manager starts, it auto-selects the
Windows driver on Windows. This change makes it
also auto-select the Support library extra package.
The rationale is that a lot of code now depends on it
(ADT templates, docs' sample code) so sooner or later
the user will need to install it anyway. We can help
newcomers by selecting it upfront.

(cherry picked from commit 15832585a03db4e9ccb0b9732c7c9fd2677f8529)

Change-Id: I26c6277deb569b58ec904342dbd98dd347c11dde