Additional fixes to filter out library project lint warnings. DO NOT MERGE
This is a cherrypick of just the bug fix portion of
https://android-review.googlesource.com/#/c/45952

Lint already has the concept of whether a project should report lint
warnings or not; a project can be included in analysis (e.g. for
unused resource checks), but can filter out any warnings local to that
project. This is useful when you are using a library, but don't want
to see errors from that library which may not be under your control.

The way this is handled from the command line is that lint will only
report errors for projects you've referenced; e.g. if you run "lint
/foo/bar" this will show errors in /foo/bar, but exclude errors found
in the library project /foo/bar/../library".

However, there were several lint checks which needed additional fixes
for this, because (like the unused error detector) they gather data
from multiple projects and process and report in the after-project
hook.

Change-Id: I4c4b3ae451866231ae03dc75e8e1c32f27a84534