Support package conflicts between app and libs. do not merge.
If 2 libraries share the same package name, this is now
a build breakage, unless all libraries share the same
package.
This is because the libraries R classes only contain the
resources declared in the libraries themselves (plus
dependencies). Since two libraries could share the same
package name without depending on one another, it's possible
to ensure that creating only one R class would work for both.
(Merging both R class might be possible but is too risky for
a quick fix like this).
If all the libraries share the same package, then a single R
class is created for that package that contains all the symbol
of the app (simpler than merging all the symbols for now)

If a library and the app share the same package name, the
R class for the library is not created (since the R class for
the app contains all resources). This already worked in ADT,
so this changeset only fixes Ant.

(cherry picked from commit 131459f779b1a98bf354663298637589a2860fff)

Change-Id: I594f817145fbc150a629cd2141ee698534717dd8