Fix target parser
The AndroidTargetParser tries to build up a view of the class hierarchy. It
does this by loading the classes and walking the class hierarchy. However,
it cannot define classes in the java.* namespace; if it encounters any of
these, it has to skip them.

Change-Id: I02c3d06f168cf3bb7e21b978160bcdc785bf09e1
(cherry picked from commit 1d869217e6d4844a4222a4ff739d0d4f4b06d919)