ADT project export: Suggest importing, and use latest versions
First, this changeset updates the gradle plugin version exported by
the wizard to pick up the "latest version" constant from tools/base
such that we don't need to update it in ADT separately (it was stuck
at version 0.5).

Second, when it copies the gradle wrapper, it simultaneously updates
the gradle wrapper distribution URL to the current minimum requirement
instead of whatever version is currently shipping in the SDK folder
where it finds the wrapper (currently 1.6).

Third, it fixes the export to look up the source level and if it's
1.7, insert compileOptions to tell Gradle to use that level as well.

Finally, it adds a new first page to the export wizard which explains
that Studio import is now available and that it tracks studio gradle
requirements more closely.

Change-Id: I8b0ec0a8e109d120dd1a7bac916e6477afac2d17
(cherry picked from commit 242f6330ea5fce287f09d25342a7e8fbd5648c1d)