Fix Dependency Project Creation
GridLayout and AppCompat are included in the SDK as dependency libraries
rather than just .jar files because they include resources.
Thus, when they are added as a dependency to a project they need to be
added as library projects to the Eclipse workspace.

There is code that automates this process, but it had a bug where
it would not insert the correct target for the library project that
it created.

This change fixes that behavior and enables the automatic library project
creation in the NewActivityWizard.

Change-Id: I18aea5b7b6b93f7ca71e98385bf00b647ea99d51