FindJava2+WinLauncher2 MFC app for Windows.
In the code base we have a "find_java" directory
that compiles using mingw32 + an Android.mk; it is
used to create the find_java.dll used by the NSIS
installer, the find_java.exe that is used by android.bat
to locate java before executing the SDK/AVD Manager
and finally also used by the Android Studio WinLauncher.

This reworks find_java to create 2 new MFC apps:
- FindJava2 is a straight replacement for the previous find_java.exe.
- WinLauncher2 will be a replacement for the Studio launcher.

The main change is that the new app has a UI so it lists
all the Java.exe paths and let the user add his/her own.
Then it remembers the selected path using a registry key
and reuses the next time if it's available.

The app is built using VS2013 and MFC and there's no
Android.mk for it. The goal is to just make prebuilts
(that's how find_java.exe/dll and WinLauncher are actually
used right now.)

The FindJava2 part is final.
What's left to do:
- Merge the code of WinLauncher into WinLauncher2 to
  actually run Studio (they will be a need for both a
  32-bit version and 64-bit version support.)
- Create a new FindJava2.dll for NSIS once we decide how
  we want to use this in the new installer (e.g. do we
  just want to make sure there is "some" version of Java
  or do we want to include the select-and-register functionality
  too so that the user doesn't have to be asked again later?)

Change-Id: I814ed46711ac17a66cd63b9e7c7d485632169ff1