emulator/opengl: Fix glShaderSource() signature.
The glShaderSource() function really takes a 'const GLchar* const*' parameter,
not 'const GLchar**'. This creates issues when compiling the auto-generated
encoders created with the latest version of 'emugen' and the GL2 Khronos
headers.

This patch is used to fix the issue. Note that it requires defining a new type
in gl2.types, as well as fixing the signature of misc functions in the
translator.

Change-Id: I750f95f2e33d83b3b5563642ad7f87e4b8a37f35
(cherry picked from commit 92a73be1a361353e0c3577002e477db68fa00d45)