emulator/opengl/emugen: Add new 'flushOnEncode' entry point flag.
This adds a new flag that can be applied to entry points for 'emugen',
named 'flushOnEncode'. When used, the generated encoder will call
stream->flush() just after adding bytes to the stream.

This is needed to match the manual change that was performed in
the renderControl encoder in the following patch:

    https://android-review.googlesource.com/#/c/95864/

Change-Id: I8cc8fdb0d38ef27e8ba646c83d717166400babfd