emulator/opengl: Fix pixelDataSize() ambiguity.
Without this fix, the GLESv1 encoder library will link to the
pixelDataSize() function that is defined in the GLESv2 encoder library,
which results in bugs, like the boot animation now showing with -gpu on.

This patch doesn't change the generated decoder libraries, nor does it
affect the encoder ones until external/qemu/distrib/update-emugl-sources.sh
is called. When the latter script will be run, other changes under
device/generic/goldfish/generic/ will be needed to fix the namespace
of the functions being defined in GLEncoderUtils.cpp and GL2EncoderUtils.cpp

Change-Id: I588659461dd6d73350d150a896494ccf50fbdc7f