find_java.bat now uses env var instead of find
The officially recommended way to query an OS for its 32-bit/64-bit
architecture can be found in this KB article:
http://support.microsoft.com/kb/556009

Unfortunately, this fails for some subset of users for at least two
reasons:

1) Many users install Unix tools on their Windows box, overwriting
   "find" with the Unix variant that doesn't support the \i parameter
2) Some machines are simply missing reg.exe. I've seen one claim in
   the wild that this was due to an agressive anti-virus scanner
   removing malware from their machine.

Therefore, we now use absolute paths to the utilities we want to
target. Additionally, we now have fallback handling if reg.exe is
missing, where we, instead, rely on the %PROCESSOR_ARCHITECTURE%
environment variable.

Related issues:
  https://code.google.com/p/android/issues/detail?id=82099
  https://code.google.com/p/android/issues/detail?id=82388

Change-Id: Icee0882fa6e1f5359c086498eb2679197cdf6f8e