SDK ProGuard config: Fix support for Parcelable.ClassLoaderCreator
This fixes the SDK ProGuard config to preserve the name of the CREATOR
field of Parcelables whose CREATOR field are Parcelable.ClassLoaderCreator
instead of the usual Parcelable.Creator.

The long description:

In the beginning, Parcel.readParcelable() worked by reading a
Parcelable's field named CREATOR and casting it to Parcelable.Creator.
This was originally supported by the ProGuard config.

Commit b46ed7636be9341b6ce0b158b3d86f34a437e6da changed
Parcel.readParcelable() to allow the Parcelable's CREATOR field to
alternatively be of type Parcelable.ClassLoaderCreator.

The problem was that the ProGuard config was the following:

-keepclassmembers class * implements android.os.Parcelable {
  public static final android.os.Parcelable$Creator CREATOR;
}

This preserved the name of a field of type Parcelable.Creator, but not
the names of fields of type Parcelable.ClassLoaderCreator.

The fix is to instead use the following:

-keepclassmembers class * implements android.os.Parcelable {
  public static final ** CREATOR;
}

For classes derived from Parcelable, this preserves the name of fields
named CREATOR of any type, which should be more resilient to any future
changes.

I tested this change by writing a small test program, confirming that a
sample Parcelable was unmarshallable at runtime, and by dumping the APK
and confirming that the field name was properly preserved.

Change-Id: I9187fbbd2745ffb0b55f6e201fd57d1420e961dc
Signed-off-by: Spencer Low <CompareAndSwap@gmail.com>