Supporting path specified by build_tools.sh
When this is run by build_tools.sh, the out path is specified
in an env var. This is the path used for output by previous steps
in the script, and this build depends on those outputs being
present. Thus currently the build will fail if the out path specified
for build_tools.sh doesn't match the hardcoded one here.
This change make this build respect the out path env var.

Change-Id: I56285c894b80c7b637c484a2cd5c9e2a59d86d5c