adb tab completion: remove extraneous state
An extraneous state caused adding -s to the command line to have an
off-by-one error when trying to tab complete the device serial number.
Instead ignore that state in the state machine but still use the state
to determine when we should try to get a serial number for completion
from adb.

Bug: 22353839
Change-Id: I0c67ed6e395ea3689ba2707981c9514c6d5edabb